(hybris) mixer_paths: Use handset mic for speaker mode workaround.
Microphone is muted with speaker route during voice calls and VOIP
calls. As a workaround use handset mic config for the speaker use case
as well. It's not nearly optimal as it introduces pretty bad echo for
the caller in as echo cancellation is not working as it should but it's
better than the alternative of no sound from microphone at all.

[hybris] mixer_paths: Use handset mic for speaker mode workaround. JB#57389

Signed-off-by: Matti Kosola <matti.kosola@jolla.com>